api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
openai==1.2.4
google-generativeai==0.2.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pypdf==3.17.0

# Django integration
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-filter==23.3
djangorestframework-simplejwt==5.3.0
asgiref>=3.6.0
